move functions to new package
